Add files via upload
four web pages with the correct links, background image and proper google font api in place. The pages are 80% complete with the header footer, font, background, links, images, content, layout completed. The css code is still currently in the html file for ease of code. Css code will be taken to its own file in the next upload.